requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
webvtt-py==0.5.1
